Network Connectivity & bug fixes

+++ app/src/main/java/com/android/ucptask/data/db/entity/CurrentWeatherEntry.kt
@@ -8,9 +8,9 @@ data class CurrentWeatherEntry(
     val feelsLike: Double,
     val humidity: Int,
     val pressure: Int,
-    val temp: Int,
+    val temp: Double,
     @SerializedName("temp_max")
-    val tempMax: Int,
+    val tempMax: Double,
     @SerializedName("temp_min")
-    val tempMin: Int
+    val tempMin: Double
 )
